Move tests to test/, make health server opt-in via config
- Move test files from src/ to test/ and update imports
- Add HEALTH_PORT config to gate health server startup
- Document all config variables in README and .env.example
- Fix test script glob for cross-platform CI compatibility
