+++ .env.example
@@ -2,8 +2,12 @@
 SLACK_BOT_TOKEN=xoxb-...
 SLACK_APP_TOKEN=xapp-...
 
-# Anthropic API key (written to auth.json on startup)
+# Anthropic credentials (provide one of these)
 ANTHROPIC_API_KEY=sk-ant-...
+# ANTHROPIC_OAUTH_REFRESH_TOKEN=...
+
+# Optional: override the default model (default: claude-sonnet-4-5)
+# MODEL=claude-sonnet-4-5
 
 # GitHub token (used by gh CLI)
 GITHUB_TOKEN=ghp_...
@@ -14,3 +18,10 @@ MAX_QUEUE_SIZE=10
 
 # Optional: Slack channel ID for audit logging bot requests
 # LOG_CHANNEL_ID=C0123456789
+
+# Optional: persist OAuth tokens in Upstash Redis
+# UPSTASH_REDIS_REST_URL=https://...
+# UPSTASH_REDIS_REST_TOKEN=...
+
+# Optional: enable health check endpoint (GET /health/live)
+# HEALTH_PORT=5000
