+++ README.md
@@ -35,7 +35,9 @@ npm start
 npm run dev
 
 # CLI mode for local testing (no Slack required)
-npm run cli
+npm run cli                              # interactive REPL
+npm run cli -- "Create an issue"         # one-shot mode
+npm run cli -- --dry-run "Describe plan" # one-shot, dry run
 ```
 
 ## Configuration
@@ -47,18 +49,26 @@ See [`.env.example`](.env.example) for all available options. Key variables:
 | `SLACK_BOT_TOKEN` | Yes | Bot token (`xoxb-...`) |
 | `SLACK_APP_TOKEN` | Yes | App-level token for Socket Mode (`xapp-...`) |
 | `GITHUB_TOKEN` | Yes | GitHub PAT for `gh` CLI |
-| `ANTHROPIC_API_KEY` | Yes | Anthropic API key |
+| `ANTHROPIC_API_KEY` | * | Anthropic API key (provide this or OAuth token) |
+| `ANTHROPIC_OAUTH_REFRESH_TOKEN` | * | OAuth refresh token (alternative to API key) |
+| `MODEL` | No | Model override (default: `claude-sonnet-4-5`) |
 | `MAX_CONCURRENT_AGENTS` | No | Max parallel agent runs (default: 3) |
 | `MAX_QUEUE_SIZE` | No | Max queued requests (default: 10) |
+| `UPSTASH_REDIS_REST_URL` | No | Upstash Redis URL for OAuth token persistence |
+| `UPSTASH_REDIS_REST_TOKEN` | No | Upstash Redis token |
+| `LOG_CHANNEL_ID` | No | Slack channel ID for audit logging |
+| `HEALTH_PORT` | No | Port for health check endpoint (`GET /health/live`) |
+
+\* One of `ANTHROPIC_API_KEY` or `ANTHROPIC_OAUTH_REFRESH_TOKEN` is required.
 
 ## Docker
 
 ```bash
 docker build -t slack-issue-bot .
-docker run -p 5000:5000 --env-file .env slack-issue-bot
+docker run --env-file .env slack-issue-bot
 ```
 
-The container exposes a health check at `GET /health/live` on port 5000.
+Set `HEALTH_PORT=5000` (and expose the port) to enable the health check endpoint.
 
 ## Project structure
 
@@ -67,10 +77,12 @@ src/
   index.ts          Entry point
   slack.ts          Slack event handlers and message processing
   agent.ts          Claude agent initialization and execution
+  prompt.ts         Prompt builder (extracted for testability)
   config.ts         Environment variable loading
   concurrency.ts    Agent scheduler with queue management
-  cli.ts            CLI interface for local testing
+  cli.ts            CLI interface for local testing (REPL + one-shot)
   health.ts         Health check endpoint
+test/               Unit tests (node:test)
 prompts/
   system.md         System prompt for the Claude agent
 skills/             Agent skill definitions (create-issue, github, repos, triage)
